bugfix: fix login failure as admin
several causes:

when DEFAULT_HOST is wrong, login attempt fails with "Can't verify CSRF
token authenticity".

as the application runs with RAILS_ENV=production, and TLS is not
enabled, session cookie will not be passed to the application, resulting
the same error.

finally, admin email and password must not be quoted. or the password
would be _"password"_ with quotes.